==> src/3_Vacation_Itinerary.ipynb <==
## 3.  Create a Travel Itinerary Map.

# Dependencies and Setup
import pandas as pd
import requests
import gmaps
import numpy as np

# Import API key
from config import g_key

# Configure gmaps
gmaps.configure(api_key=g_key)

# Read the WeatherPy_vacation.csv into a DataFrame.
vacation_df = pd.read_csv("resources/WeatherPy_vacation.csv")
vacation_df.head()

# Add the city name, the country code, the weather description and maximum temperature for the city.
info_box_template = """
<dl>
<dt>Hotel Name</dt><dd>{Hotel Name}</dd>
<dt>City</dt><dd>{City}</dd>
<dt>Country</dt><dd>{Country}</dd>
<dt>Current Weather</dt><dd>{Current Description} and {Max Temp} °F</dd>
</dl>
"""
clean_hotel_df = vacation_df.dropna()
# Get the data from each row and add it to the formatting template and store the data in a list.
hotel_info = [info_box_template.format(**row) for index, row in clean_hotel_df.iterrows()]

# Get the latitude and longitude from each row and store in a new DataFrame.
locations = clean_hotel_df[["Lat", "Lng"]]

# Add a marker layer for each city to the map.
max_temp = clean_hotel_df["Max Temp"]
fig = gmaps.figure(center=(30.0, 31.0), zoom_level=1.5)
marker_layer = gmaps.marker_layer(locations, info_box_content=hotel_info)
fig.add_layer(marker_layer)
fig


# create a vacation route using 4 cities from map above

# Te Anay, NZ
vacation_start = vacation_df.loc[vacation_df["City"]=="Te Anau"]
# Tuatapere, NZ
vacation_stop1 = vacation_df.loc[vacation_df["City"]=="Tuatapere"]
# Mataura, NZ
vacation_stop2 = vacation_df.loc[vacation_df["City"]=="Mataura"] 
# Kaitangata, NZ
vacation_stop3 = vacation_df.loc[vacation_df["City"]=="Kaitangata"]

# Get the latitude-longitude pairs
start = vacation_start[["Lat","Lng"]].to_numpy()[0]
stop1 = vacation_stop1[["Lat","Lng"]].to_numpy()[0]
stop2 = vacation_stop2[["Lat","Lng"]].to_numpy()[0]
stop3 = vacation_stop3[["Lat","Lng"]].to_numpy()[0]


# Create a direction layer map 
fig = gmaps.figure()
route_map = gmaps.directions_layer(
        start, start, waypoints=[stop1,stop2,stop3],
        travel_mode='DRIVING')
fig.add_layer(route_map)
fig


# Create a marker layer map between the four cities.
itinerary_df = pd.concat([vacation_start, vacation_stop1, vacation_stop2, vacation_stop3],ignore_index=True)
itinerary_df

# Add city name, the country code, the weather description and maximum temperature for the city. 
info_box_template = """
<dl>
<dt>Hotel Name</dt><dd>{Hotel Name}</dd>
<dt>City</dt><dd>{City}</dd>
<dt>Country</dt><dd>{Country}</dd>
<dt>Current Weather</dt><dd>{Current Description} and {Max Temp} °F</dd>
</dl>
"""

# Add data from each row to formatting template and store in list
hotel_info = [info_box_template.format(**row) for index, row in itinerary_df.iterrows()]

# store lat long for each city in new DF
locations = itinerary_df[["Lat", "Lng"]]

# Add a marker layer for each city to the map.
max_temp = itinerary_df["Max Temp"]
fig = gmaps.figure(center=(30.0, 31.0), zoom_level=1.5)
marker_layer = gmaps.marker_layer(locations, info_box_content=hotel_info)
fig.add_layer(marker_layer)
fig
